pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "numpy", "cython"]
build-backend = "setuptools.build_meta"

[project]
name = "cama-flood-python"
version = "4.2.0"
description = "CaMa-Flood洪水模拟模型Python重构版本"
authors = [{name = "CaMa-Flood Team"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "numba>=0.56.0",
    "xarray>=0.20.0",
    "netcdf4>=1.5.0",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "matplotlib>=3.5.0",
    "cartopy>=0.20.0",
    "dask>=2021.10.0",
    "pytest>=8.3.5",
]

[project.optional-dependencies]
dev = ["pytest>=6.0", "pytest-cov", "black", "flake8"]
viz = ["plotly>=5.0", "bokeh>=2.4"]
hpc = ["mpi4py>=3.1", "h5py>=3.6"]

[project.scripts]
cama-flood = "cama_flood_py.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cama_flood_py*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
